[runtime] Rename llvm/build.mk fragment, because now automake
Originally this was a bit of Make. We moved it up and were
including it as an automake blob. This broke because you can't
just include a Makefile from an automake file. The syntaxes differ.

I added some stub logic that would forward the targets we wanted
to the build.mk makefile, so the Makefile generated from the llvm/Makefile.am file
would correctly interact with the fragment.

Instead, we seem to want to use the automake fragment directly and
include it. Ludovic rewrote the makefile syntax into the subset of make
that is also valid automake syntax.

Because this is an automake file, I'd like to rename it to build.am.
Otherwise future contributors might mistake this for a make fragment and add Make syntax.
This now needs automake syntax, or the shared subset of syntax that is
valid in each.